+++ mobile/android/android-components/components/browser/engine-gecko-nightly/src/main/java/mozilla/components/browser/engine/gecko/TrackingProtectionExceptionFileStorage.kt
@@ -103,8 +103,13 @@ internal class TrackingProtectionExceptionFileStorage(
         persist()
     }
 
-    override fun removeAll() {
+    override fun removeAll(activeSessions: List<EngineSession?>?) {
         runtime.contentBlockingController.clearExceptionList()
+        activeSessions?.forEach {
+            it?.notifyObservers {
+                onExcludedOnTrackingProtectionChange(false)
+            }
+        }
         removeFileFromDisk(context)
     }